% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumprocess.R
\name{crit.SBQ.Rtest}
\alias{crit.SBQ.Rtest}
\title{Critical values for the retrospective stacked backward CUSUM test}
\usage{
crit.SBQ.Rtest(k, alternative = "two.sided")
}
\arguments{
\item{k}{Number of regressors and dimension of the CUSUM process}

\item{alternative}{A character string specifying the alternative hypothesis; must be one of "two.sided" (default) or "one.sided".}
}
\value{
A vector of critical values for different significance levels
}
\description{
Provides critical values for some selected values for k for the stacked backward CUSUM test with the linear boundary \eqn{d(r) = 1+2r} for different significance levels.
}
\examples{
crit.SBQ.Rtest(1)
}
