% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumprocess.R
\name{get.crit.SBQ.mon}
\alias{get.crit.SBQ.mon}
\title{Critical values for the stacked backward CUSUM monitorng procedure}
\usage{
get.crit.SBQ.mon(k, m = Inf, alternative = "two.sided")
}
\arguments{
\item{k}{Number of regressors and dimension of the CUSUM process}

\item{m}{The length of the relative monitoring period; m = Inf for infinite horizon monitoring (default).}

\item{alternative}{A character string specifying the alternative hypothesis; must be one of "two.sided" (default) or "one.sided".}
}
\value{
A list containing the following components:
\item{critical.values}{A vector of critical values for different significance levels}
\item{m}{The value for the length of the relative monitoring period m. If the critical values for the input value of m is not available, the next higher available value is considered.}
}
\description{
Provides critical values for some selected combinations of k and m for the stacked backward CUSUM monitoring procedure with the linear boundary \eqn{d(r) = 1+2r} for different significance levels.
}
\examples{
get.crit.SBQ.mon(1,Inf)
get.crit.SBQ.mon(3,0.5)
}
