% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumprocess.R
\name{get.partialcusum}
\alias{get.partialcusum}
\title{Partial CUSUM process}
\usage{
get.partialcusum(formula, T, H)
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}

\item{T}{In the retropspective context: length of the sample. In the monitoring context: length of the training sample, where monitoring starts at T+1.}

\item{H}{A matrix for the partial hypothesis \eqn{H'\beta_t = H'\beta_0}. \eqn{H} must have orthonormal columns.}
}
\value{
A matrix containing the partial forward CUSUM process \eqn{Q_T^*(r)}
}
\description{
Partial CUSUM process
}
\examples{
T <- 100
u <- rnorm(T,0,1)
x <- rnorm(T,0,1)
y <- c(rep(0,T/2), rep(1,T/2)) + x + u
H <- matrix(c(1,0), ncol = 1)
get.partialcusum(y~x, T, H=H)
}
