% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors_retrospective.R
\name{get.recresid}
\alias{get.recresid}
\title{Recursive residuals}
\usage{
get.recresid(formula)
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}
}
\value{
A vector containing the recursive residuals
}
\description{
Recursive residuals
}
\examples{
T <- 100
u <- rnorm(T,0,1)
y <- c(rep(0,T/2), rep(1,T/2)) + u
get.recresid(y~1)
}
