% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors_monitoring.R
\name{Q.mon.csw}
\alias{Q.mon.csw}
\title{CUSUM monitoring by Chu, Stinchcombe, and White (1996)}
\usage{
Q.mon.csw(formula, T, alpha = 0.05, alternative = "two.sided")
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}

\item{T}{Length of the training sample. Monitoring starts at T+1.}

\item{alpha}{Significance level of the monitoring procedure (alpha = 0.05 is default).}

\item{alternative}{A character string specifying the alternative hypothesis; must be one of "two.sided" (default), "greater" or "less".
The detector statistic is given by the maximum norm of \eqn{Q_t} ("two.sided"), maximum entry of \eqn{Q_t} ("greater"), or maximum entry of \eqn{-Q_t} ("less"), respectively.}
}
\value{
A list containing the following components:
\item{detector}{The vector containing the path of the forward cusum detector from T+1 onwards}
\item{boundary}{The vector containing the values of the boundary function from T+1 onwards}
\item{rejection}{A logical vector containing the test decision for different significance levels; TRUE for rejection}
\item{detectontime}{The vector containing the detection time point, which is the time indices of the first boundary crossing; NA if the null hypothesis is not rejected}
\item{statistic}{The test statistic; maximum of the detector scaled by the boundary function}
}
\description{
Performs the univariate forward CUSUM monitoring procedure using the radical boundary of Chu, Stinchcombe, and White (1996) for infinite horizon monitoring.
}
\examples{
T <- 100
t <- 5*T
u <- rnorm(t,0,1)
x <- rnorm(t,1,2)
y <- c(rep(0,480), rep(5,20)) + x + I(x^2) + u
Q.mon.csw(y~1+x+I(x^2), T)
Q.mon.csw(y~1+x+I(x^2), T, alternative = "greater")
}
