% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors_monitoring.R
\name{Q.mon.lin}
\alias{Q.mon.lin}
\title{Multivariate forward CUSUM monitoring}
\usage{
Q.mon.lin(formula, T, m = Inf, alternative = "two.sided", H = NULL)
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}

\item{T}{Length of the training sample. Monitoring starts at T+1.}

\item{m}{The length of the relative monitoring period; default is m = Inf (infinite horizon monitoring).}

\item{alternative}{A character string specifying the alternative hypothesis; must be one of "two.sided" (default), "greater" or "less".
The detector statistic is given by the maximum norm of \eqn{Q_t} ("two.sided"), maximum entry of \eqn{Q_t} ("greater"), or maximum entry of \eqn{-Q_t} ("less"), respectively.}

\item{H}{An optional matrix for the partial hypothesis \eqn{H'\beta_t = H'\beta_0}, where \eqn{H'Q_t} is considered instead of \eqn{Q_t}.
\eqn{H} must have orthonormal columns. The full structural break test is considered as the default setting (NULL).}
}
\value{
A list containung the following components:
\item{detector}{The vector containing the path of the forward cusum detector from T+1 onwards}
\item{boundary}{The vector containing the values of the boundary function from T+1 onwards}
\item{critical.value}{A vector containing critical values for different significance levels; NA if critical value for this specification is not implemented}
\item{rejection}{A logical vector containing the test decision for different significance levels; TRUE for rejection; NA if critical value is not implemented}
\item{detectontime}{The vector containing the detection time points for different significance levels, which are the time indices of the first boundary crossing; NA if the null hypothesis is not rejected}
\item{statistic}{The test statistic; maximum of the detector scaled by its boundary \eqn{d(r)}}
}
\description{
Performs the multivariate forward CUSUM monitoring procedure using the linear boundary of Brown, Durbin, and Evans (1975)
}
\examples{
T <- 100
t <- 5*T
u <- rnorm(t,0,1)
x <- rnorm(t,1,2)
y <- c(rep(0,480), rep(5,20)) + x + I(x^2) + u
Q.mon.lin(y~1+x+I(x^2), T)
Q.mon.lin(y~1+x+I(x^2), T, alternative = "greater")
H <- matrix(c(1,0,0), ncol = 1)
Q.mon.lin(y~1+x+I(x^2), T, m=6, H = H)
}
