% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors_monitoring.R
\name{Qmon.detector}
\alias{Qmon.detector}
\title{CUSUM monitoring detector}
\usage{
Qmon.detector(formula, T, bound = "linear", alternative = "two.sided")
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}

\item{T}{Length of the training sample. Monitoring starts at T+1.}

\item{bound}{Type of the boundary. Either 'linear' or 'radical'}
}
\value{
A vector containing the Forward CUSUM monitoring detector series
}
\description{
CUSUM monitoring detector
}
\examples{
T <- 100
u <- rnorm(T,0,1)
y <- c(rep(0,T/2), rep(1,T/2)) + u
Q.detector(y~1)
}
