% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors_retrospective.R
\name{SBQ.test}
\alias{SBQ.test}
\title{Stacked backward CUSUM R-test (retrospective)}
\usage{
SBQ.test(formula, alternative = c("two.sided", "greater", "less"), H = NULL)
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}

\item{alternative}{A character string specifying the alternative hypothesis; must be one of "two.sided" (default), "greater" or "less".
The output detector matrix are the maximum norms of \eqn{SBQ_{s,t}} ("two.sided"), the maximum entries of \eqn{SBQ_{s,t}}("greater"), or maximum entries of \eqn{-SBQ_{s,t}} ("less"), respectively.}

\item{H}{An optional matrix for the partial hypothesis \eqn{H'\beta_t = H'\beta_0}, where \eqn{H'BQ_t} is considered instead of \eqn{SBQ_{s,t}}.
\eqn{H} must have orthonormal columns. For a test for a break in the intercept, H can also set to the string "intercept".
The full structural break test is considered as the default setting (NULL).}
}
\value{
A list containing the following components:
\item{detector.array}{A matrix containing the triangular array of the stacked backward cusum detector depending on the specificaton for the alternative hypothesis}
\item{boundary}{The matrix containing the values of the triangular boundary surface}
\item{detector.scaled}{A vector containing the path of the sequential stacked backward cusum detector divided by the boundary}
\item{statistic}{The test statistic; maximum of detector.scaled}
\item{alternative}{The specification for the alternative hypothesis}
\item{critical.value}{A vector containing critical values for different significance levels; NA if critical value for this specification is not implemented}
\item{rejection}{A logical vector containing the test decision for different significance levels; TRUE for rejection; NA if critical value is not implemented}
}
\description{
Performs the multivariate stacked backward CUSUM R-test (retrospective) with linear boundary
}
\examples{
T <- 100
u <- rnorm(T,0,1)
x <- rnorm(T,1,2)
y <- c(rep(0,T/2), rep(0.7,T/2)) + x + I(x^2) + u
SBQ.test(y~1+x+I(x^2))
SBQ.test(y~1+x+I(x^2), alternative = "greater")
H <- matrix(c(1,0,0), ncol = 1)
SBQ.test(y~1+x+I(x^2), H = H)
}
