% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors_monitoring.R
\name{SBQmon.detector}
\alias{SBQmon.detector}
\title{Stacked backward CUSUM detector}
\usage{
SBQmon.detector(formula, T, alternative = "two.sided")
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}

\item{T}{Length of the training sample. Monitoring starts at T+1.}
}
\value{
A vector containing the Stacked backward CUSUM monitoring detector series
}
\description{
Stacked backward CUSUM detector
}
\examples{
T <- 100
u <- rnorm(T,0,1)
y <- c(rep(0,T/2), rep(1,T/2)) + u
BQ.detector(y~1)

}
