% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumprocess.R
\name{sim.crit.SBQ}
\alias{sim.crit.SBQ}
\title{Simulates critical values for the multivariate stacked backward CUSUM test}
\usage{
sim.crit.SBQ(
  k,
  m = 2,
  levels = c(0.1, 0.05, 0.01, 0.001),
  alternative = "two.sided",
  MC = 100,
  T = 1000
)
}
\arguments{
\item{k}{Number of regressors and dimension of the CUSUM process.}

\item{m}{The length of the relative monitoring period; m = Inf for infinite horizon monitoring; for retrospective testing set m = 2 (default).}

\item{levels}{A vector if significance levels.}

\item{alternative}{A character string specifying the alternative hypothesis; must be one of "two.sided" (default) or "one.sided".}

\item{MC}{Number of Monte Carlo replications.}

\item{T}{gridsize.}
}
\value{
A vector of critical values for different significance levels.
}
\description{
The asymptotic critical values for the linear boundary \eqn{d(r) = 1+2r} are simulated based on \eqn{MC} Monte Carlo replications, where the limiting process under the null hypothesis is approximated on a grid of \eqn{T} equidistant points.
}
\examples{
sim.crit.Q(1)
}
