% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors_retrospective.R
\name{sup.wald}
\alias{sup.wald}
\title{Sup-Wald Test by Andrews (1993)}
\usage{
sup.wald(formula, eps = 0.15)
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}

\item{eps}{Trimming parameter, eps = 0.15 is default.}
}
\value{
The sup-Wald test statistic
}
\description{
Performs the Sup-Wald Test by Andrews (1993).
}
\examples{
T <- 100
u <- rnorm(T,0,1)
x <- rnorm(T,1,2)
y <- c(rep(0,T/2), rep(0.7,T/2)) + x + I(x^2) + u
sup.wald(y~1+x+I(x^2))
}
