% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumprocess.R
\name{get.cusumprocess}
\alias{get.cusumprocess}
\title{Multivariate CUSUM process \eqn{Q_T(r)}}
\usage{
get.cusumprocess(formula, T = NULL)
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}

\item{T}{Optional length of the training sample for monitoring. For retrospective tests, T should be the sample size. If NULL, T is stet to the sample size. NULL is default.}
}
\value{
A matrix containing the multivariate forward CUSUM process \eqn{Q_T(r)}
}
\description{
Multivariate CUSUM process \eqn{Q_T(r)}
}
\examples{
T <- 100
u <- rnorm(T,0,1)
x <- rnorm(T,0,1)
y <- c(rep(0,T/2), rep(1,T/2)) + x + u
get.cusumprocess(y~x, T)
}
