% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumprocess.R
\name{get.partialcusum}
\alias{get.partialcusum}
\title{Partial CUSUM process \eqn{Q_T^*(r)}}
\usage{
get.partialcusum(formula, T = NULL, H = "intercept")
}
\arguments{
\item{formula}{Specification of the linear regression model by an object of the class "formula"}

\item{T}{Optional length of the training sample for monitoring. For retrospective tests, T should be the sample size. If NULL, T is stet to the sample size. NULL is default.}

\item{H}{An optional matrix for the partial hypothesis \eqn{H'\beta_t = H'\beta_0}. \eqn{H} must have orthonormal columns.
If "intercept", H is set to the setting of testing for a break only in the first regressor variable (typically the intercept). "intercept" is default.}
}
\value{
A matrix containing the partial forward CUSUM process \eqn{Q_T^*(r)}
}
\description{
Partial CUSUM process \eqn{Q_T^*(r)}
}
\examples{
T <- 100
u <- rnorm(T,0,1)
x <- rnorm(T,0,1)
y <- c(rep(0,T/2), rep(1,T/2)) + x + u
H <- matrix(c(1,0), ncol = 1)
get.partialcusum(y~x, T, H=H)
}
